pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "wcs-analyzer"
version = "0.1.0"
description = "WCS dance video analyzer with WSDC-style competition scoring"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "opencv-python>=4.9.0",
    "librosa>=0.10.0",
    "ffmpeg-python>=0.2.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.26.0",
]

[project.scripts]
wcs-analyzer = "wcs_analyzer.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/wcs_analyzer"]
